Moved PortType to its own file
