+++ lluvia/cpp/core/include/lluvia/core.h
@@ -32,6 +32,7 @@
 #include "core/node/NodeState.h"
 #include "core/node/NodeType.h"
 #include "core/node/PortDirection.h"
+#include "core/node/PortType.h"
 
 #include "core/CommandBuffer.h"
 #include "core/ComputeNode.h"
