+++ lluvia/python/src/lluvia/core/enums/__init__.py
@@ -7,6 +7,5 @@
 """
 
 from .compute_dimension import ComputeDimension
-from .node import PortType
 from .vulkan import Format, ImageType
 from .parameter import ParameterType
