+++ lluvia/python/src/lluvia/core/node.pxd
@@ -21,12 +21,9 @@ from lluvia.core.types cimport _vec3ui
 
 from lluvia.core.node_new.node_type cimport _NodeType
 from lluvia.core.node_new.port_direction cimport _PortDirection
+from lluvia.core.node_new.port_type cimport _PortType
 
 cdef extern from 'lluvia/core/Node.h' namespace 'll':
-
-    cdef cppclass _PortType 'll::PortType':
-        pass
-
 
     cdef struct _PortDescriptor 'll::PortDescriptor':
         uint32_t binding
